pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qnn-mnist"
version = "0.1.0"
description = "Hybrid Quantum Neural Network for MNIST classification"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Dishant Gupta", email = "22bph016@nith.ac.in"}
]
dependencies = [
    "pennylane>=0.39.0",
    "pennylane-lightning>=0.39.0",
    "torch>=2.3.1",
    "torchvision>=0.18.1",
    "scikit-learn>=1.5.1",
    "numpy>=1.26.4",
    "matplotlib>=3.9.0",
    "streamlit>=1.37.1",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.2",
    "black>=24.0.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"